+++ samplePDFDUNE/samplePDFDUNEBaseND.h
@@ -49,33 +49,33 @@ struct dunendmc_base {
 #endif
 
   // histo pdf bins
-  double *rw_erec;
-  double *rw_etru;
-  double *rw_elep_reco;
-  double *rw_mom;
-  double *rw_theta;
-  double *rw_Q2;
-  double *rw_yrec;
-  double *rw_pdf_bin_1d; //global bin in the PDF so we can do a faster fill1Dhist 
-  double *rw_lower_erec_1d; // lower to check if Eb has moved the erec bin
-  double *rw_upper_erec_1d; // upper to check if Eb has moved the erec bin
-  double *rw_pdf_bin_2d; //global bin in the PDF so we can do a faster fill2Dhist 
-  double *rw_lower_erec_2d; // lower to check if Eb has moved the erec bin
-  double *rw_upper_erec_2d; // upper to check if Eb has moved the erec bin
+  float *rw_erec;
+  float *rw_etru;
+  float *rw_elep_reco;
+  float *rw_mom;
+  float *rw_theta;
+  float *rw_Q2;
+  float *rw_yrec;
+  float *rw_pdf_bin_1d; //global bin in the PDF so we can do a faster fill1Dhist 
+  float *rw_lower_erec_1d; // lower to check if Eb has moved the erec bin
+  float *rw_upper_erec_1d; // upper to check if Eb has moved the erec bin
+  float *rw_pdf_bin_2d; //global bin in the PDF so we can do a faster fill2Dhist 
+  float *rw_lower_erec_2d; // lower to check if Eb has moved the erec bin
+  float *rw_upper_erec_2d; // upper to check if Eb has moved the erec bin
   int *rw_reco_nue;
   int *rw_reco_numu;
-  double *rw_berpaacvwgt;
+  float *rw_berpaacvwgt;
   int    *rw_isCC;
   int    *rw_nuPDGunosc;
   int    *rw_nuPDG;
   int    *rw_run;
   bool    *rw_isFHC;
-  double *rw_vtx_x;
-  double *rw_vtx_y;
-  double *rw_vtx_z;
-  double dummy_y;
+  float *rw_vtx_x;
+  float *rw_vtx_y;
+  float *rw_vtx_z;
+  float dummy_y;
 
-  double *reco_numu;
+  float *reco_numu;
 
   float **rw_dirlep;
   int *mode;
@@ -83,13 +83,13 @@ struct dunendmc_base {
   int **rw_ipnu;
 
 
-  double pot_s; // s is for scale                                             
-  double norm_s;//    
+  float pot_s; // s is for scale                                             
+  float norm_s;//    
 
-  double *beam_w;
-  double *flux_w; // not the same as beam systematics weight!                 
-  double *xsec_w;
-  double *energyscale_w;
+  float *beam_w;
+  float *flux_w; // not the same as beam systematics weight!                 
+  float *xsec_w;
+  float *energyscale_w;
   //float *relRPA_w;
 
 };
@@ -142,7 +142,7 @@ public:
   TTree *_data;
   TString _nutype;
   int _mode;
-  float _pnu[50];
+  double _pnu[50];
   double _wgtflx;
   double _wgtosc;
   int _ipnu;
